requirements.txt
bidict==0.23.1
blinker==1.9.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-SocketIO==5.6.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.0
python-barcode==0.16.1
python-engineio==4.13.0
python-socketio==5.16.0
reportlab==4.4.9
simple-websocket==1.1.0
Werkzeug==3.1.5
wsproto==1.3.2